// planning-standards-system/src/common/rbac/permission.enum.d.ts
export declare enum Permission {
    SERVICES_READ = "services.read",
    SERVICES_WRITE = "services.write",
    KPIS_READ = "kpis.read",
    KPIS_WRITE = "kpis.write",
    HOLIDAYS_READ = "holidays.read",
    HOLIDAYS_WRITE = "holidays.write",
    PERIODS_READ = "periods.read",
    PERIODS_WRITE = "periods.write",
    COMMITMENTS_READ = "commitments.read",
    COMMITMENTS_WRITE = "commitments.write",
    COMMITMENTS_LOCK = "commitments.lock"
}